fix: create training attention mask in f32 to avoid bf16 overflow
torch.finfo(torch.float32).min overflows when masked_fill_ targets
a bf16 tensor. Create in f32 first, then cast to bf16.

+++ modelopt/torch/speculative/plugins/hf_dflash.py
@@ -706,9 +706,9 @@ class HFDFlashModel(DFlashModel):
 
         # Convert bool mask to float additive mask for SDPA
         dtype = target_hidden.dtype
-        attn_mask = torch.zeros(bsz, 1, q_len, kv_len, device=device, dtype=dtype)
+        attn_mask = torch.zeros(bsz, 1, q_len, kv_len, device=device, dtype=torch.float32)
         attn_mask.masked_fill_(~final_mask, torch.finfo(torch.float32).min)
-        attn_mask = attn_mask.to(dtype)
+        attn_mask = attn_mask.to(dtype=dtype)
 
         # 7. Draft forward
         hidden = self.dflash_module(
